+++ packages/plugin-react-app/CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## 2.0.4
 
+- [fix] error occurred when config externals and remoteRuntime
 - [fix] build error when active `remoteRuntime`
 
 ## 2.0.3